Moved ObjectProcessorQueue to queues, added some doc
